Refactor: Lowercase component folders, to fix warnings about 'modules with names that only differ in casing'.
